# backend-node/openapi.yaml
openapi: 3.0.3
info:
  title: AI Services Platform API
  version: 2.0.0
  description: APIs for AI-powered services including Virtual Assistant, IDP, Computer Vision, Authentication, and more.
  contact:
    name: AI Services Platform Support
    email: support@aiservicesplatform.com
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://api.yourdomain.com
    description: Production server
...
paths:
...
  /api/usage/assistant-call:
    post:
      summary: Update usage metrics for a call or session
      tags: [Usage]
      description: Receives usage metrics from Java VA service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                callId:
                  type: string
                  description: Call ID (ObjectId) or session ID (UUID)
                customerId:
                  type: string
                sttSeconds:
                  type: number
                ttsCharacters:
                  type: number
                llmTokensIn:
                  type: integer
                llmTokensOut:
                  type: integer
                provider:
                  type: string
                costEstimate:
                  type: number
              required: [callId]
      responses:
        '200':
          description: Usage updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  callId:
                    type: string
        '404':
          description: Call not found
  /api/usage/assistant-call/{callId}:
    get:
      summary: Get usage metrics for a specific call
      tags: [Usage]
      parameters:
        - in: path
          name: callId
          required: true
          schema:
            type: string
          description: Call ID (ObjectId) or session ID (UUID)
      responses:
        '200':
          description: Usage metrics
          content:
            application/json:
              schema:
                type: object
                properties:
                  callId:
                    type: string
                  customerId:
                    type: string
                  usage:
                    type: object
                    properties:
                      sttSeconds:
                        type: number
                      ttsCharacters:
                        type: number
                      llmTokensIn:
                        type: integer
                      llmTokensOut:
                        type: integer
        '404':
          description: Call not found

  # Transactions Endpoints
  /api/transactions:
    get:
      summary: Get all transactions for tenant
      tags: [Transactions]
      security:
        - cookieAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
        - in: query
          name: status
          schema:
            type: string
            enum: [success, failed, pending, refunded]
        - in: query
          name: paymentMethodId
          schema:
            type: string
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '401':
          description: Unauthorized
  /api/transactions/{transactionId}:
    get:
      summary: Get transaction by ID
      tags: [Transactions]
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: transactionId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transaction details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  transaction:
                    $ref: '#/components/schemas/Transaction'
        '404':
          description: Transaction not found
  /api/transactions/payment-method/{paymentMethodId}:
    get:
      summary: Get transactions for a specific payment method
      tags: [Transactions]
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: paymentMethodId
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
  /api/transactions/stats/summary:
    get:
      summary: Get transaction statistics
      tags: [Transactions]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Transaction statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  stats:
                    type: object
                    properties:
                      total:
                        type: integer
                      successful:
                        type: integer
                      failed:
                        type: integer
                      pending:
                        type: integer
                      refunded:
                        type: integer
                      totalAmount:
                        type: number
                      averageAmount:
                        type: number
                      thisMonth:
                        type: number
                      lastMonth:
                        type: number

  # Tenants Endpoints
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session cookie for authentication
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Transaction:
      type: object
      properties:
        _id:
          type: string
        tenantId:
          type: string
        customerId:
          type: string
        paymentMethodId:
          type: string
        amount:
          type: number
          description: Amount in cents
        currency:
          type: string
          default: USD
        status:
          type: string
          enum: [success, failed, pending, refunded]
        description:
          type: string
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...